requirements.txt
discord.py>=2.3.0
asyncpg>=0.28.0
redis>=5.0.0
python-dotenv>=1.0.0
psutil>=5.8.0
aiofiles>=23.0.0
